refactor: pass email date params to worker and cron
